Added comment and file header. Changed long int to long long int

+++ pyccel/stdlib/numpy/numpy_c.c
@@ -1,6 +1,12 @@
+/* --------------------------------------------------------------------------------------- */
+/* This file is part of Pyccel which is released under MIT License. See the LICENSE file   */
+/* or go to https://github.com/pyccel/pyccel/blob/master/LICENSE for full license details. */
+/* --------------------------------------------------------------------------------------- */
+
 #include "numpy_c.h"
 
-long int isign(long int x)
+/* numpy.sign for float, double and integers */
+long long int isign(long long int x)
 {
     return SIGN(x);
 }
